header update
ok this one is kind of huge, i finally was able to get the avionics view to work with 3 badges in the top slider menu and not look like dogshit on my mobile.  i'm pretty happy about this update ;-), now I need to apply it to satcom and astrometrics as those just look awful currently on mobile devices.